client update 로비 기능 구현 1.방 목록 받아와서 셋팅하기 2.방 생성하기 -> 게임방 입장하기
할것
1.채팅
2.페이지 버튼 눌렀을때 페이지 정보 받아오는거
3.섹터매니저 틀 구현

+++ Team.Yggdrasil_Server/test/Assets/after/NetWork/NetObjectInfo.cs
@@ -17,10 +17,10 @@ namespace Net
             Spirit,
             Monster,
         }
-        [SerializeField] ulong m_id;
+        [SerializeField] uint m_id;
         ENetObjectType m_type;
 
-        public ulong GetID { get => m_id; }
+        public uint GetID { get => m_id; }
         public ENetObjectType GetNetObjectType { get => m_type; }
         public NetObjectInfo(ENetObjectType _type)
         {
